Bug Fix in custom_load related to OpenGL

+++ Paris-Lille3D/KPConv/NZPC_KPConv_Inference.py
@@ -181,8 +181,8 @@ def main():
         
         vis_d = {
             "name": batch['name'],
-            "points": batch['point'],
-            "labels": batch['label'],
+            "points": batch['point'].astype(np.float32),
+            "labels": batch['label'].astype(np.float32),
             "pred": pred_label_k,
                 }
